% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_plot.R
\name{boundary_plot}
\alias{boundary_plot}
\title{Convience function for creating boundary plots}
\usage{
boundary_plot(model, x, y, z, xlab = all.vars(formula(model))[2],
  ylab = all.vars(formula(model))[3], levels = 0.5, lwd = 2,
  jitter = TRUE, ...)
}
\arguments{
\item{model}{A caret classification model}

\item{x}{Data}

\item{y}{Data}

\item{z}{Observed values}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{levels}{contour levels, default prob=0.5}

\item{lwd}{Line width of the contour plot}

\item{jitter}{Default \code{TRUE}. Should the points be jittered.}

\item{...}{Additional arguments passed to the contour function}
}
\description{
Convience function for creating boundary plots
}
