% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concrete.R
\docType{data}
\name{concrete}
\alias{concrete}
\title{The concrete mixture data set}
\value{
A data.frame
}
\description{
Measurements of proportions of ingredients in concrete composition.
The aim is to be able to predict the compressive strength of the 
concrete based on its composition. This is a modified version of
a data set (mixtures) in the AppliedPredictiveModelling package
}
\examples{
data(concrete)
}
