% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrPredictive-package.R
\docType{package}
\name{jrPredictive-package}
\alias{jrPredictive-package}
\alias{jrPredictive}
\title{Predictive analytics}
\description{
Functions and datasets used for the predictive analytics course by Jumping Rivers.
}
\keyword{package}
