% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark}
\alias{mark}
\title{A function for marking models during the predictive analytics course}
\usage{
mark(model)
}
\arguments{
\item{model}{A model resulting from a call to train}
}
\description{
A function for marking models during the predictive analytics course
}
