% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortgage.R
\docType{data}
\name{mortgage}
\alias{mortgage}
\title{The mortgage data set}
\value{
A data.frame
}
\description{
1049 measurements of 16 variables from a US bank. 
The goal is to be able to predict the 30 year mortgage rate,
X30YCMortgageRate.
}
\examples{
data(mortgage)
}
