% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{A function for validating models for submission during the predictive analytics course}
\usage{
validate(model = NULL)
}
\arguments{
\item{model}{A model resulting from a call to train}
}
\value{
A model of type train
}
\description{
A function for validating models for submission during the predictive analytics course
}
